#1539: Kubernetes and plain docker compatible resolvers

+++ ndla-run-kong.sh
@@ -1,18 +1,25 @@
 #!/bin/sh
 
+function is_kubernetes {
+    [ -e "/var/run/secrets/kubernetes.io" ]
+}
+
 function prepare_remote {
-    secretsfile="/tmp/secrets"
-    aws s3 --region eu-central-1 cp s3://$NDLA_ENVIRONMENT.secrets.ndla/api-gateway.secrets $secretsfile
+    if is_kubernetes; then
+        export KONG_CLUSTER_ADVERTISE=$(hostname -i):7946
+    else
+        secretsfile="/tmp/secrets"
+        aws s3 --region eu-central-1 cp s3://$NDLA_ENVIRONMENT.secrets.ndla/api-gateway.secrets $secretsfile
 
-    export KONG_CLUSTER_ADVERTISE=$HOST_IP:7946
+        export KONG_CLUSTER_ADVERTISE=$HOST_IP:7946
 
-    export KONG_PG_HOST=$(cat $secretsfile | jq -r .META_SERVER)
-    export KONG_PG_PORT=$(cat $secretsfile | jq -r .META_PORT)
-    export KONG_PG_DATABASE=$(cat $secretsfile | jq -r .META_RESOURCE)
-    export KONG_PG_USER=$(cat $secretsfile | jq -r .META_USER_NAME)
-    export KONG_PG_PASSWORD=$(cat $secretsfile | jq -r .META_PASSWORD)
-
-    rm $secretsfile
+        export KONG_PG_HOST=$(cat $secretsfile | jq -r .META_SERVER)
+        export KONG_PG_PORT=$(cat $secretsfile | jq -r .META_PORT)
+        export KONG_PG_DATABASE=$(cat $secretsfile | jq -r .META_RESOURCE)
+        export KONG_PG_USER=$(cat $secretsfile | jq -r .META_USER_NAME)
+        export KONG_PG_PASSWORD=$(cat $secretsfile | jq -r .META_PASSWORD)
+        rm $secretsfile
+    fi
 }
 
 function setup_logging {
@@ -36,6 +43,14 @@ function setup_nginx_caches {
     fi
 }
 
+function setup_dns_resolver {
+    if is_kubernetes; then # Check whether we are running on kubernetes or not
+        echo "resolver kube-dns.kube-system.svc.cluster.local;" > /nginx-resolver.conf
+    else
+        echo "resolver 127.0.0.11;" > /nginx-resolver.conf
+    fi
+}
+
 if [ "$NDLA_ENVIRONMENT" != "local" ]
 then
     prepare_remote
@@ -43,6 +58,7 @@ fi
 
 setup_logging
 setup_nginx_caches
+setup_dns_resolver
 
 export KONG_PROXY_LISTEN=0.0.0.0:8000
 export KONG_ADMIN_LISTEN=0.0.0.0:8001
